Fix tests and update comments
